Fix segfault in drawFrame()
There was a segfault caused by maxFramesInFlight being set to less than
the max number of swapchain images. This segfault only occasionally
occurred, and I think this was because sometimes the swapchain would be
set to double buffering, instead of triple buffering, which changed the
max number of swapchain images to be greater than maxFramesInFlight.

Also unindented code in main().