Move drawFrame and swapchain lifecycle to Engine
This commit moves drawFrame(), cleanupSwapchain(), and
recreateSwapchain() from their respective namespaces to the Engine
class. The motivation for this was that 20+ Engine state members were
being passed to these functions, so it made more sense to merge them in.
This also has the benefit of removing 40+ line function declarations,
and huge function calls.