Add Phong shader, light class, and normal maps
Phong lighting was implemented in the fragment shader. The model on the
left has hard coded normals, the model in the middle uses an uncorrected
normal map, and the model on the right displays the normal map as
colors. Next step: Implement the TBN matrix to apply normal mapping in
tangent space, allowing normals to be correct even after model rotates.

A point light class was created which currently supports specifying its
color, and 'patroling' between two specified points. Implmenting this
revealed a reliance on undefined behavior in the camera class, where its
yaw and pitch member variables were uninitialized. Creating an instance
of the light class caused the memory in the address of the uninitialized
pitch member variable to hold a 'garbage' value, preventing the camera
from pointing horizontally from mouse input.

Also, the alignas specifier had to be used to add several glm::vec3's to
the UniformBufferObject struct, since the Vulkan expects data sent to a
shader in a UBO to be aligned in certain ways.